+++ texar/modules/classifiers/gpt2_classifier.py
@@ -62,7 +62,7 @@ class GPT2Classifier(ClassifierBase):
                  cache_dir: Optional[str] = None,
                  hparams=None):
 
-        super().__init__(hparams)
+        super().__init__(hparams=hparams)
 
         # Create the underlying encoder
         encoder_hparams = dict_fetch(hparams, GPT2Encoder.default_hparams())
